pyproject.toml
[tool.poetry]
name = "fast-curso"
version = "0.1.0"
description = ""
authors = ["GabMeula07 <gabrielqueirozmeula@gmail.com>"]
readme = "README.md"

[tool.taskipy.tasks]
run = 'fastapi dev fast_curso/app.py'
pre_test = 'task lint'
test = 'pytest --cov=fast_curso -vv'
post_test = 'coverage html'
lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . --fix ; ruff format'

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
ruff = "^0.5.5"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies] 
ruff = "^0.5.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.pytest.ini_options]
pythonpath="."
addopts = '-p no:warnign'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
